Can register users and login
Okay many gradle changes to update but thats ok.

Redid the whole login activities to 2
A login and register

Remember when you add activities to add them to the manifest file

Now there we firebase code errors but fixed. There can be bugs but idk yet

We are using firebase Authentication. which means you can use email to reset password or 2 factor login.

The only downside is that we are using firebase Authentication which means its not "connected" to the database because it's saving the users in a specific area, so we will have to figure out how to move them there. A cool feature about it is that we will be able to use emails to reset passwords.

The way I set up the app is that MainActivity loads 1st. This will be the Splash screen but gotta fix the wait function

Then LogInActivity loads next. If registering the RegisterActivity loads then goes back to LogInActivity

After sign in the user will enter MainPageActvity